dialects: (test) add test.allocatable and use in unit tests
